requirements.txt
python-dotenv
requests
pefile
signify